covalence-hol: split out a top-level `category` module
Move the `Category` trait and the commutative-diagram API out of
`monoidal/` into their own top-level `category` module, so the categorical
foundation is independent of the coproduct's symmetric-monoidal structure.
This is the intended home for the string-diagram calculus (a category, not
a coproduct, is all string diagrams need).

- `category::Category` — the trait (was `monoidal::category`).
- `category::diagram` — commutative diagrams (was `monoidal::diagram`).
- `category::shallow::Hol` — the shallow HOL model now implements
  `Category` here, forwarding to `init::cat`; `category` depends only on
  `init::cat`, not on `monoidal` or `init::coprod`.

`monoidal` now builds on top: `Monoidal: category::Category`,
`impl Monoidal for Hol` stays in `monoidal::shallow`, and `monoidal`
re-exports `Category`/`Hol` so existing paths keep working. Add a
`concat` path test. Tests, clippy, and rustdoc are clean across both
modules.
