+++ crates/covalence-hol/src/lib.rs
@@ -19,6 +19,7 @@
 //! Nothing in this crate is consumed by `covalence-core`'s inference
 //! rules.
 
+pub mod category;
 pub mod hash;
 pub mod hol_light_obs;
 pub mod init;
